% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acceptAllEffects.R
\name{acceptAllEffects}
\alias{acceptAllEffects}
\title{Accepts all estimates for fixed effects, Sigma, and random effects}
\usage{
acceptAllEffects(model)
}
\arguments{
\item{model}{Model object}
}
\value{
\code{NlmePmlModel} object
}
\description{
Updates PML statements in model object with estimates returned from model execution. Use \code{\link{copyModel}} and set
argument \code{acceptAllEffects = TRUE} to create new model object with final estimates from base model run.
}
\examples{
\donttest{
model <- acceptAllEffects(model)
}
}
\seealso{
\code{\link{copyModel}}
}
\keyword{internal}
