% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{addExtraDef}
\alias{addExtraDef}
\title{Adds user defined extra column/table definitions to column definition file}
\usage{
addExtraDef(.Object, value)
}
\arguments{
\item{.Object}{PK/PD model}

\item{value}{Character vector of extra column/table definitions}
}
\value{
Modified \code{NlmePmlModel} object
}
\description{
Adds user defined extra column/table definitions to column definition file
}
\examples{
\donttest{
model <- addExtraDef(model, c("addlcol(ADDL)", "table(file=\"res.csv\",time(0),Ka,V,Cl,Tlg)"))
}

}
