% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dosing.r
\name{addMDV}
\alias{addMDV}
\title{Adds MDV extra column definition to model object}
\usage{
addMDV(.Object, MDV)
}
\arguments{
\item{.Object}{Model object}

\item{MDV}{Column mapping argument specifying corresponding "MDV" column in input data set}
}
\value{
Modified \code{NlmePmlModel} object
}
\description{
Use to add MDV statement to \code{model@userDefinedExtraDefs}
}
\examples{
\donttest{
model <- addMDV(model, MDV = "MDV")
}
}
