% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{addReset}
\alias{addReset}
\alias{addReset,NlmePmlModel-method}
\title{Adds reset instructions to the model}
\usage{
addReset(.Object, low, hi, Reset = NULL)

\S4method{addReset}{NlmePmlModel}(.Object, low, hi, Reset = NULL)
}
\arguments{
\item{.Object}{An 'NlmePmlModel' object to which you want to add reset instructions.}

\item{low}{Lower value of reset range.}

\item{hi}{Upper value of reset range.}

\item{Reset}{Name of reset column in input data set for column mapping.
The default is NULL.}
}
\value{
Depends on the specific methods

Returns the 'NlmePmlModel' object with updated reset information and definitions.
}
\description{
Adds reset instructions to the model
}
\section{Functions}{
\itemize{
\item \code{addReset(NlmePmlModel)}: Method for the 'NlmePmlModel' class

This method adds reset instructions to the NlmePmlModel object.
It updates the reset information, checks column mappings if input data is not null,
and adds a reset definition to user-defined extra definitions.

}}
