% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTablesToColumnMapping.R
\name{addTablesToColumnMapping}
\alias{addTablesToColumnMapping}
\title{Add tables to the mapping}
\usage{
addTablesToColumnMapping(model, Tables, filename, forSim = TRUE)
}
\arguments{
\item{model}{Model object}

\item{Tables}{Tables class objects e.g., \code{NlmeTableDef} or \code{NlmeSimTableDef}}

\item{filename}{Name of column definition file}

\item{forSim}{Logical; Set to \code{TRUE} if simulation tables are provided}
}
\value{
Table class objects specified to \code{Tables} argument.
}
\description{
Add tables to the mapping
}
\examples{
\donttest{
addTablesToColumnMapping(model, simParams, Tables, filename)
}

}
\keyword{internal}
