% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_model.r
\name{addToErrorModel}
\alias{addToErrorModel}
\title{Add to the NLME Error model}
\usage{
addToErrorModel(model, effectsList)
}
\arguments{
\item{model}{Model object}

\item{effectsList}{List of effects}
}
\description{
Add to the NLME Error model
}
\keyword{internal}
