% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootstrapParams.R
\docType{class}
\name{BootstrapParams-class}
\alias{BootstrapParams-class}
\alias{BootstrapParams}
\title{Represents parameters for a bootstrap run}
\description{
Class initializer for BootstrapParams that represents parameters for a bootstrap run.
}
\section{Slots}{

\describe{
\item{\code{numReplicates}}{Number of bootstrap replicates to run}

\item{\code{initialEstimates}}{Get model final estimates to run bootstrap (T|F)}

\item{\code{numRetries}}{Number of times to retry a failed replicated}

\item{\code{randomNumSeed}}{Seed for random number generator}

\item{\code{confidenceLevel}}{Confidence level}

\item{\code{stratifyColumns}}{What columns to stratify on(comma separated)}
}}

\examples{
# same object as BootstrapParams()
boot <- BootstrapParams(
  numReplicates = 100,
  initialEstimates = FALSE,
  numRetries = 2,
  randomNumSeed = 1234,
  confidenceLevel = 95
)

}
\keyword{internal}
