% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleNlmeJob.R
\name{cancelJob}
\alias{cancelJob}
\alias{cancelJob,SimpleNlmeJob-method}
\title{Generic function for cancelling a job}
\usage{
cancelJob(.Object)

\S4method{cancelJob}{SimpleNlmeJob}(.Object)
}
\arguments{
\item{.Object}{A 'SimpleNlmeJob' object that you want to cancel}
}
\value{
Depends on the specific methods

Prints the 'SimpleNlmeJob' object after attempting to cancel the job.
No return value.
}
\description{
Generic function for cancelling a job
}
\section{Functions}{
\itemize{
\item \code{cancelJob(SimpleNlmeJob)}: Method for cancelling a job of the 'SimpleNlmeJob' class

This method attempts to cancel a job of the 'SimpleNlmeJob' class.
If the job is running on a local host or is not running in the background,
it throws an error and does nothing. Otherwise, it uploads a 'STOP' command
to the host's remote executor.

}}
