% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkHostParams.R
\name{checkHostParams}
\alias{checkHostParams}
\title{Check Host Parameters}
\usage{
checkHostParams(obj, verbose = FALSE)
}
\arguments{
\item{obj}{NLME Parallel Host to be checked}
}
\value{
\code{TRUE} if all checks are successful, otherwise \code{FALSE}
}
\description{
Checks NLME Parallel Host object for correct settings for GCC, NLME Installation, MPI and Root directories.
}
\examples{
\dontrun{

# multicore
multicoreHost <- hostParams(
  parallelMethod = "Multicore",
  hostName = "local_multicore",
  numCores = 4
)
checkHostParams(multicoreHost)
}

}
\keyword{NLME}
\keyword{internal}
