% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{covariateEffect<-}
\alias{covariateEffect<-}
\title{Sets style for a covariate/variable}
\usage{
covariateEffect(.Object, covariateName, parameterName) <- value
}
\arguments{
\item{.Object}{A PK/PD model}

\item{covariateName}{Name of the covariate}

\item{parameterName}{Name of the model variable}

\item{value}{A value to set}
}
\description{
Sets style for a covariate/variable
}
\keyword{internal}
