% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovariateEffectModel.R
\docType{class}
\name{CovariateEffectModel-class}
\alias{CovariateEffectModel-class}
\alias{CovariateEffectModel}
\title{NLME covariate effects model object class}
\description{
Class represents an NLME covariate effects model
}
\section{Slots}{

\describe{
\item{\code{numCovariates}}{Number of covariate effects}

\item{\code{covariateList}}{Comma separated list of covariate effects names}

\item{\code{scenarioNames}}{Comma separated list of scenario names}

\item{\code{isDefault}}{Comma separated list of flags}

\item{\code{degreesOfFreedom}}{Comma separated list of degrees of freedom}
}}

\examples{
CovariateEffectModel(numCovariates = 2,
                     covariateList = "V-Age,Cl-BW",
                     scenarioNames = "S,S",
                     isDefault = "1,1",
                     degreesOfFreedom = "1,1")
}
\keyword{NLME}
\keyword{internal}
