% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariateNames.R
\name{covariateNames}
\alias{covariateNames}
\title{Return covariate names}
\usage{
covariateNames(model)
}
\arguments{
\item{model}{Model object}
}
\value{
Character vector of covariate names defined in model
}
\description{
Use to return character vector of covariate names available in model object.
}
\examples{
model <- pkmodel(columnMap = FALSE)
model <- addCovariate(model, covariate = "BW", effect = "V")
model <- addCovariate(model, covariate = "Age", effect = "Cl")

covariateNames(model)

}
