% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NlmeCovariateParameter.r
\name{covariatePartsString}
\alias{covariatePartsString}
\title{Creates string from each covariate attribute}
\usage{
covariatePartsString(obj)
}
\arguments{
\item{obj}{Model covariate}
}
\description{
Creates string from each covariate attribute
}
\keyword{internal}
