% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model_info.R
\name{createModelInfo}
\alias{createModelInfo}
\title{Parse the model and get the list of terms}
\usage{
createModelInfo(model, ForceRun = FALSE)
}
\arguments{
\item{model}{Model object}

\item{ForceRun}{Set to \code{TRUE} to force run}
}
\value{
List of model information
}
\description{
Calls TDL5 to parse the model and get the list of terms
}
\examples{
\donttest{
  model <- pkmodel(columnMap = FALSE,
                   workingDir = tempdir())
  createModelInfo(model)
}

}
