% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NlmeColumnMapping.r
\name{dataMapping}
\alias{dataMapping}
\title{Initialize input data for PK/PD model}
\usage{
dataMapping(.Object, data)
}
\arguments{
\item{.Object}{Model object}

\item{data}{Input data of class \code{data.frame}.}
}
\value{
Modified \code{NlmePmlModel} object
}
\description{
Used to initialize input data for PK/PD model
}
\examples{
model <- pkmodel(columnMap = FALSE,
                 workingDir = tempdir())

model <- dataMapping(model, pkData)

}
\seealso{
\code{\link{colMapping}}
}
