% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{deleteSecondary}
\alias{deleteSecondary}
\alias{deleteSecondary,NlmePmlModel-method}
\title{Deletes a secondary parameter from the model}
\usage{
deleteSecondary(.Object, name)

\S4method{deleteSecondary}{NlmePmlModel}(.Object, name)
}
\arguments{
\item{.Object}{An 'NlmePmlModel' object from which you want to delete a secondary parameter.}

\item{name}{Name of the secondary parameter to be deleted.}
}
\value{
Depends on the specific methods

Returns the 'NlmePmlModel' object with the secondary parameter removed.
}
\description{
Deletes a secondary parameter from the model
}
\section{Functions}{
\itemize{
\item \code{deleteSecondary(NlmePmlModel)}: Method for the 'NlmePmlModel' class

This method deletes a secondary parameter from the NlmePmlModel object.
It searches for the parameter by name, and if it is found, it removes it
from the object and updates the PML model.

}}
\keyword{internal}
