% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dosing.r
\name{doseNames}
\alias{doseNames}
\title{Return dose names}
\usage{
doseNames(model)
}
\arguments{
\item{model}{Model object}
}
\value{
Character vector of dose names defined in model
}
\description{
Use to return character vector of dose point names in model object.
}
\examples{
model <- pkmodel(columnMap = FALSE)

doses <- doseNames(model)

}
