% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editModel.R
\name{editModel}
\alias{editModel}
\title{Directly edit PML text in model object}
\usage{
editModel(.Object)
}
\arguments{
\item{.Object}{Model object}
}
\value{
Modified \code{NlmePmlModel} object
}
\description{
Allows user to edit PML text in model object using internal text editor and return a new textual model
containing the edited PML statements.
}
\examples{
\donttest{
model <- pkmodel(columnMap = FALSE)

newModel <- editModel(model)
}

}
