% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dosing.r
\docType{class}
\name{ExtraDoseItem-class}
\alias{ExtraDoseItem-class}
\alias{ExtraDoseItem}
\title{Defines an extra dose point}
\description{
Defines an extra dose point
}
\section{Slots}{

\describe{
\item{\code{name}}{Dose point name}

\item{\code{type}}{Character; Options are "Bolus" or "Infusion"}

\item{\code{amount}}{Optional; Character specifying corresponding column in data or numeric specifying dose amount}

\item{\code{rate}}{Optional; Character specifying corresponding column in data or numeric specifying dose rate}

\item{\code{deltaTime}}{Optional; Character specifying corresponding column in data or numeric specifying delta time}

\item{\code{isSecondDose}}{Use second dose point on compartment?}

\item{\code{data}}{Optional data frame. Must specify \code{data} if supplying column as character value to \code{amount, rate, deltaTime} arguments}
}}

\keyword{internal}
