% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{generatePML}
\alias{generatePML}
\title{Generates PML statements based on the current model}
\usage{
generatePML(.Object)
}
\arguments{
\item{.Object}{PK/PD Model}
}
\description{
Generates PML statements based on the current model
}
\keyword{internal}
