% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{generatePMLModel}
\alias{generatePMLModel}
\title{Generates PML statements based on the current model}
\usage{
generatePMLModel(.Object)
}
\arguments{
\item{.Object}{PK/PD model}
}
\description{
Generates PML statements based on the current model
}
\keyword{internal}
