% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NlmeRandomEffectBlock.r
\name{getRandomEffectNames}
\alias{getRandomEffectNames}
\title{Return random effect names in model}
\usage{
getRandomEffectNames(model)
}
\arguments{
\item{model}{Model object}
}
\value{
Characters vector of random effect names
}
\description{
Use to return character vector of random effect names (if available) in model object
}
\examples{
model <- pkmodel(columnMap = FALSE)
getRandomEffectNames(model)

}
