% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getThetas.R
\name{getThetas}
\alias{getThetas}
\title{Return theta names and values}
\usage{
getThetas(model)
}
\arguments{
\item{model}{PK/PD model}
}
\value{
Character vector of theta names defined in model
}
\description{
Returns named character vector of theta values by parsing PML fixed effect statements
}
\examples{
\donttest{
getThetas(pkpdmodel)
}
}
