% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomEffectsMethods.R
\name{initializeRandomEffectsBlock}
\alias{initializeRandomEffectsBlock}
\title{Initializes random effects structure from structural parameters}
\usage{
initializeRandomEffectsBlock(.Object)
}
\arguments{
\item{.Object}{PK/PD model}
}
\description{
Initializes random effects structure from structural parameters
}
\keyword{internal}
