% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearmodel.R
\name{linearmodel_MappingParameters}
\alias{linearmodel_MappingParameters}
\title{Linear model mapping parameters}
\usage{
linearmodel_MappingParameters(ID = NULL, C = NULL, EObs = NULL)
}
\arguments{
\item{ID}{Column mapping argument for input dataset column(s)
that identify individual data profiles. Only applicable to population models
\code{isPopulation = TRUE}.}

\item{C}{Column mapping argument that represents
the input dataset column for the independent variable that is treated
as a covariate during the estimation/simulation process.}

\item{EObs}{Column mapping argument that represents
the input dataset column for the observed drug effect
(i.e., the dependent variable).}
}
\description{
Linear model mapping parameters
}
\keyword{internal}
