% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{listCovariateEffectNames}
\alias{listCovariateEffectNames}
\alias{listCovariateEffectNames,NlmePmlModel-method}
\title{Lists covariate effect names in the model}
\usage{
listCovariateEffectNames(.Object)

\S4method{listCovariateEffectNames}{NlmePmlModel}(.Object)
}
\arguments{
\item{.Object}{PK/PD model}
}
\value{
A vector of character strings containing the names
of the covariate effects in the model.
}
\description{
This function lists the names of covariate effects in a provided
pharmacokinetic/pharmacodynamic (PK/PD) model.
}
\examples{
\donttest{
listCovariateEffectNames(model)
}
}
