% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveModel.R
\name{loadModel}
\alias{loadModel}
\title{Load model.rda file}
\usage{
loadModel(directory)
}
\arguments{
\item{directory}{Directory where the model was saved}
}
\value{
No value is returned. Model object is loaded in global environment.
}
\description{
Loads a previously saved model from disk.
}
\examples{
\donttest{
loadModel(directory)
}

}
\keyword{internal}
