% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelVariableNames.R
\name{modelVariableNames}
\alias{modelVariableNames}
\title{Return model variable names}
\usage{
modelVariableNames(model)
}
\arguments{
\item{model}{Model object}
}
\value{
Character vector of required model variable names
}
\description{
Return a vector of model variable names from model object
}
\examples{
\donttest{
model <- pkmodel(columnMap = FALSE,
                 workingDir = tempdir())
modelVariableNames(model)
}

}
