% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NlmeColumnMapping.r
\docType{class}
\name{NlmeColumnMap-class}
\alias{NlmeColumnMap-class}
\alias{NlmeColumnMap}
\title{Class represents map between a model variable and a dataset column}
\arguments{
\item{variableName}{Model variable name}

\item{columnName}{Corresponding column name in the dataset}

\item{variableType}{Model variable type}
}
\description{
Class represents map between a model variable and a dataset column
}
\keyword{internal}
