% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NlmeColumnMapping.r
\docType{class}
\name{NlmeColumnMapping-class}
\alias{NlmeColumnMapping-class}
\alias{NlmeColumnMapping}
\title{Class represents mapping list between model variables and dataset columns}
\description{
Class represents mapping list between model variables and dataset columns
}
\keyword{internal}
