% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NlmeDataset.r
\docType{class}
\name{NlmeDataset-class}
\alias{NlmeDataset-class}
\alias{NlmeDataset}
\title{NLME dataset object}
\arguments{
\item{dataFile}{Subject datafile}

\item{colDefFile}{Datafile to model mapping file}

\item{modelFile}{PML model file}

\item{estimatesDefFile}{Initial estimates mapping file}

\item{estimatesDataFile}{Initial estimates values}

\item{doseDefFile}{Dose column definition file}

\item{doseDataFile}{Dose datafile}

\item{ranEffectDefFile}{Random effects column definition file}

\item{ranEffectDataFile}{Random effects data file}

\item{predoutFilename}{Name of predcheck output file}

\item{phoenixSourceDir}{Directory containing phoenix generated files}

\item{workingDir}{Directory containing datafiles(default cwd)}
}
\description{
Class represents an NLME dataset object
}
\examples{
dataset <- NlmeDataset(workingDir)
}
\keyword{internal}
