% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NlmeDoseMapping.R
\docType{class}
\name{NlmeDoseMapping-class}
\alias{NlmeDoseMapping-class}
\alias{NlmeDoseMapping}
\title{Class represents mapping list between model variables and dose columns}
\description{
Class represents mapping list between model variables and dose columns
}
\keyword{internal}
