% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_params.r
\name{NlmeEngineExtraParams}
\alias{NlmeEngineExtraParams}
\title{Create a new NlmeEngineExtraParams object}
\usage{
NlmeEngineExtraParams(...)
}
\arguments{
\item{...}{Named arguments to override the default values.}
}
\value{
An object of class NlmeEngineExtraParams.
}
\description{
This function creates a new instance of the NlmeEngineExtraParams class
}
\keyword{internal}
