% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_model.r
\docType{class}
\name{NlmeErrorModel-class}
\alias{NlmeErrorModel-class}
\alias{NlmeErrorModel}
\title{Class represents an NLME/PML error model}
\arguments{
\item{effectsList}{List of residual effects to include in the error model}

\item{numberOfEffects}{Number of effects being included in the error model}
}
\description{
Class represents an NLME/PML error model
}
\keyword{internal}
