% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NlmeIndirectParameters.R
\docType{class}
\name{NlmeIndirectParameters-class}
\alias{NlmeIndirectParameters-class}
\alias{NlmeIndirectParameters}
\title{Class represents an NLME/PML Indirect PD model parameters}
\arguments{
\item{type}{Indirect model type}

\item{hasEffectsCompartment}{Is there data available for an effects compartment?}

\item{isBuildup}{Is the response formation (TRUE) or
degradation (FALSE) concentration dependent?}

\item{isExponent}{Is there an exponent in the effect statement?}

\item{frozen}{Freeze standard deviation to prevent estimation
of the PK part of the model}
}
\description{
Class represents an NLME/PML Indirect PD model parameters
}
\keyword{internal}
