% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NlmeModelAbsorption.R
\docType{class}
\name{NlmeModelAbsorption-class}
\alias{NlmeModelAbsorption-class}
\alias{NlmeModelAbsorption}
\title{Class represents an NLME/PML model absorption}
\arguments{
\item{absorpType}{Intravenous|Extravascular}
}
\description{
Class represents an NLME/PML model absorption
}
\examples{
\donttest{
NlmeModelAbsorption(PARAM_EXTRAVASCULAR)
}
}
\keyword{internal}
