% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NlmeModelParameterization.R
\docType{class}
\name{NlmeModelParameterization-class}
\alias{NlmeModelParameterization-class}
\alias{NlmeModelParameterization}
\title{Class represents an NLME/PML model parameterization}
\arguments{
\item{paramType}{One of Micro|Clearance|Macro|Macro1}
}
\description{
Class represents an NLME/PML model parameterization
}
\examples{
\donttest{
NlmeModelParameterization(PARAM_CLEARANCE)
}
}
\keyword{internal}
