% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NlmeModelType.R
\docType{class}
\name{NlmeModelType-class}
\alias{NlmeModelType-class}
\alias{NlmeModelType}
\title{Class represents an NLME/PML model type}
\arguments{
\item{modelType}{Model type as numeric value from \code{1:7}}
}
\description{
Class represents an NLME/PML model type
}
\examples{
\donttest{
NlmeModelType(PK)
}
}
\keyword{internal}
