% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NlmeParallelMethod.R
\docType{class}
\name{NlmeParallelMethod-class}
\alias{NlmeParallelMethod-class}
\alias{NlmeParallelMethod}
\title{Class initializer for NlmeParallelMethod}
\description{
Use to define NlmeParallelMethod as for \code{parallelMethod} argument in
\code{\link[=NlmeParallelHost]{NlmeParallelHost()}}.
}
\section{Slots}{

\describe{
\item{\code{method}}{Options are:
\code{None|Multicore|LOCAL_MPI|SGE|SGE_MPI|TORQUE|TORQUE_MPI|LSF|LSF_MPI|SLURM|SLURM_MPI}.}
}}

\keyword{internal}
