% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NlmeParamsMapping.R
\docType{class}
\name{NlmeParamsMapping-class}
\alias{NlmeParamsMapping-class}
\alias{NlmeParamsMapping}
\title{Class represents mapping list between model variables and data columns}
\description{
Class represents mapping list between model variables and data columns
}
\keyword{internal}
