% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NlmeRandomEffectBlock.r
\docType{class}
\name{NlmeRandomEffectBlock}
\alias{NlmeRandomEffectBlock}
\title{Random effect block definition}
\description{
Random effect block definition
}
\section{Slots}{

\describe{
\item{\code{type}}{Diagonal or Block}

\item{\code{effectNames}}{Character or character vector specifying names of random effects}

\item{\code{frozen}}{Logical; Are random effects frozen?}
}}

\examples{
\donttest{
NlmeRandomEffectBlock(Diagonal, list("nCl", "nV"), FALSE)
}

}
\keyword{internal}
