% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NlmeRandomEffectBlock.r
\docType{class}
\name{NlmeRandomEffectValues-class}
\alias{NlmeRandomEffectValues-class}
\alias{NlmeRandomEffectValues}
\title{Random effect values matrix}
\arguments{
\item{numEffects}{Number of random effects}

\item{effectNames}{Names of random effects}

\item{values}{Matrix of random effect values}
}
\description{
Random effect values matrix
}
\examples{
\donttest{
# initialize with default effectValues ()
NlmeRandomEffectValues(list("nCl", "nV"))
# set diagonals
NlmeRandomEffectValues(list("Cl", "V"), effectValues = c(0.1, 0.2))
# values matrix overrides effectValues
NlmeRandomEffectValues(list("Cl", "V"),
  values = matrix(c(1,0.01,0.01,1), ncol = 2,
    dimnames = list(c("Cl", "V"), c("Cl", "V"))))
}
}
\keyword{internal}
