% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NlmeRandParamsMapping.R
\docType{class}
\name{NlmeRandParamsMapping-class}
\alias{NlmeRandParamsMapping-class}
\alias{NlmeRandParamsMapping}
\title{Class represents mapping list between model variables and Random Effect}
\description{
Class represents mapping list between model variables and Random Effect
}
\keyword{internal}
