% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NlmeRemoteExecutor.R
\docType{class}
\name{NlmeRemoteExecutor-class}
\alias{NlmeRemoteExecutor-class}
\alias{NlmeRemoteExecutor}
\title{Class initializer for NlmeRemoteExecutor}
\description{
Creates Remote Executor object class
}
\section{Slots}{

\describe{
\item{\code{sharedDirectory}}{The directory in which the run executes}

\item{\code{installationDirectory}}{Directory containing NLME libraries/scripts}

\item{\code{hostName}}{IP or name of remote host}

\item{\code{userAuthentication}}{Credential for user to log into remote system}
}}

\keyword{internal}
