% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NlmeScenario.R
\docType{class}
\name{NlmeScenario-class}
\alias{NlmeScenario-class}
\alias{NlmeScenario}
\title{Class initializer for NlmeScenario}
\arguments{
\item{scenarioName}{Name of the scenario}

\item{covariatesList}{Comma separated indices of covariate effects to use for this scenario}
}
\description{
Creates NlmeScenario class object
}
\examples{
CovariateEffectNames <- c("dVdBodyWeight",  "dCldBodyWeight")
#   Do not use any covariate effects
scenario1 <- NlmeScenario("no_covariates", "")
#   Use covariate effect index 1
scenario2 <- NlmeScenario("dVdBodyWeight", "1")
#   Use second covariate effect in the model
scenario3 <- NlmeScenario("dCldBodyWeight", "2")
#   Use 1st and 2nd covariate effect
scenario4 <- NlmeScenario("dVdBodyWeight_dCldBodyWeight", "1,2")
}
\keyword{internal}
