% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NlmeTableDef.R
\docType{class}
\name{NlmeSimTableDef-class}
\alias{NlmeSimTableDef-class}
\alias{NlmeSimTableDef}
\title{Class initializer for NlmeSimTableDef}
\arguments{
\item{name}{Name of the generated simulation file.}

\item{timesList}{Numeric; Time values for simulation.
Applicable for time-based models only. Ignored when \code{"keepSource=TRUE"}}

\item{covrSet}{Character; Vector of covariate names. Simulation point is added
when the covariate value is set. See \code{\link{covariateNames}}}

\item{whenDose}{Character; Vector of dosing compartment names.
Simulation point is added when the dose value is set.}

\item{whenObs}{Character; String of observed variables names.
Simulation point is added when the observation value is set.}

\item{variablesList}{Character; List of variables from the model for simulation.}

\item{keepSource}{Logical; Set to \code{TRUE} to keep the number of rows appearing
in the table the same as the number of rows in the input dataset.}

\item{timeAfterDose}{Set to \code{TRUE} to output time after dose.}
}
\description{
Creates NlmeSimTableDef class object used to specify parameters for VPC/Simulation runs
}
\keyword{internal}
