% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NlmeUserAuthentication.R
\docType{class}
\name{NlmeUserAuthentication-class}
\alias{NlmeUserAuthentication-class}
\alias{NlmeUserAuthentication}
\title{Class initializer for NlmeUserAuthentication}
\description{
Use for authentication records
}
\section{Slots}{

\describe{
\item{\code{userName}}{How the user is identified to the remote system}

\item{\code{privateKeyFile}}{path to private key file, see \code{\link[ssh:ssh_connect]{keyfile}}
for details}

\item{\code{userPassword}}{either a string or a callback function for password prompt, see
\code{\link[ssh:ssh_connect]{passwd}} for details}
}}

\keyword{internal}
