% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterNames.R
\name{observationNames}
\alias{observationNames}
\title{Get observation names}
\usage{
observationNames(model)
}
\arguments{
\item{model}{Model object}
}
\value{
Character vector of observation model variables names.
}
\description{
Get observation model variables names.
}
\examples{
model <- pkemaxmodel(columnMap = FALSE)
obsnames <- observationNames(model)

}
\keyword{internal}
