% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{OneCpt_IVInfusionData}
\alias{OneCpt_IVInfusionData}
\title{Pharmacokinetic dataset containing 100 subjects with single dose given by infusion}
\format{
A data frame with 800 rows and 6 variables:
\describe{
  \item{Subject}{Subject ID}
  \item{Time}{Time point}
  \item{Dose}{Amount of dose}
  \item{CObs}{Observations of drug concentration in blood}
  \item{Rate}{Rate of infusion}
  \item{Duration}{Duration of infusion}
}
}
\source{
The data is simulated using a PK model described by a one-compartment model with IV infusion
}
\usage{
OneCpt_IVInfusionData
}
\description{
Pharmacokinetic dataset containing 16 subjects with single dose given by infusion.
}
\keyword{datasets}
