% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_metamodel.R
\name{parse_NLMEHosts}
\alias{parse_NLMEHosts}
\title{Create NlmeParallelHost object from json file with host definition}
\usage{
parse_NLMEHosts(nlme_hostPath)
}
\arguments{
\item{nlme_hostPath}{json file with host definition for model execution.
See \code{NlmeParallelHost} class definition.
If \code{missing}, MPI local host with 4 threads is used for simple estimation
mode, multicore host is used for the others.}
}
\value{
the \code{NlmeParallelHost} class object is returned.
}
\description{
Create NlmeParallelHost object from json file with host definition
}
\examples{
\donttest{
# path nlme_hostPath should be specified

host <- parse_NLMEHosts(nlme_hostPath)
}
}
\keyword{internal}
