% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsePMLColMap.R
\name{parsePMLColMap}
\alias{parsePMLColMap}
\title{Embed column definition info into the model}
\usage{
parsePMLColMap(.Object, ForceRun = TRUE)
}
\arguments{
\item{.Object}{Model (\code{NlmePmlModel}) object}

\item{ForceRun}{Set to \code{TRUE} to force run}
}
\value{
modified \code{NLMEPmlModel} object with column mapping
definitions
}
\description{
Add/update column definition information for the model object
}
\details{
Intended to be used by other packages
}
