% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{pkData}
\alias{pkData}
\title{Pharmacokinetic dataset containing 16 subjects with single bolus dose}
\format{
A data frame with 112 rows and 8 variables:
\describe{
  \item{Subject}{Subject ID}
  \item{Nom_Time}{Nominal Time}
  \item{Act_Time}{Actual Time}
  \item{Amount}{Amount of dose}
  \item{Conc}{Observations of drug concentration in blood}
  \item{Age}{Age}
  \item{BodyWeight}{Body weight}
  \item{Gender}{Gender ("male", "female")}
}
}
\source{
\href{https://www.certara.com/training/}{Certara University}
}
\usage{
pkData
}
\description{
Pharmacokinetic dataset containing 16 subjects with single bolus dose.
}
\keyword{datasets}
