% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{pkpdData}
\alias{pkpdData}
\title{Pharmacokinetic/Pharmacodynamic dataset containing 200 subjects with single bolus dose}
\format{
A data frame with 2600 rows and 5 variables:
\describe{
  \item{ID}{Subject ID}
  \item{Time}{Nominal Time}
  \item{Dose}{Amount of dose}
  \item{CObs}{Observations of drug concentration in blood}
  \item{EObs}{Observations of drug effect}
}
}
\source{
The data is simulated using a PKPD model with PK described by a one-compartment model with IV bolus and PD described by an indirect response model with the loss inhibited.
}
\usage{
pkpdData
}
\description{
Pharmacokinetic/Pharmacodynamic dataset containing 200 subjects with single bolus dose.
}
\keyword{datasets}
