% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{print.NlmePmlModel}
\alias{print.NlmePmlModel}
\title{Print generic for class NlmePmlModel}
\usage{
\method{print}{NlmePmlModel}(x, ...)
}
\arguments{
\item{x}{NlmePmlModel class instance}

\item{...}{Arguments passed to methods.}
}
\value{
\code{NULL}
}
\description{
Prints model information, including PML and column mappings.
}
\examples{
model <- pkmodel(columnMap = FALSE, data = pkData)
print(model)

}
