% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleNlmeJob.R
\name{print.SimpleNlmeJob}
\alias{print.SimpleNlmeJob}
\title{Print generic for class Simple.NlmeJob}
\usage{
\method{print}{SimpleNlmeJob}(x, ...)
}
\arguments{
\item{x}{Handle to an NLME job}

\item{...}{Arguments passed to methods.}
}
\value{
\code{NULL}
}
\description{
Reads progress file and prints out its contents
}
\examples{
\donttest{
  print(jobHandle)
}
}
\keyword{internal}
