% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProfileParameters.R
\docType{class}
\name{ProfileParameters-class}
\alias{ProfileParameters-class}
\alias{ProfileParameters}
\title{Class initializer for ProfileParameters}
\description{
Class represents an NLME profile perturbation variable
}
\section{Slots}{

\describe{
\item{\code{howToPertubate}}{How to apply profile variables. Options are \code{USE_DELTA} or \code{USE_PERCENTAGE}}

\item{\code{profileVars}}{List of profile variables}
}}

\examples{
profile1 <- ProfileVar("tvV", 9.548, "-2,-1,0,1,2")
profile2 <- ProfileVar("tvCl", 3.219, "-1,0,1")
profiles <- ProfileParameters("USE_DELTA", c(profile1, profile2))
}
\keyword{internal}
