% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProfileVar.R
\docType{class}
\name{ProfileVar-class}
\alias{ProfileVar-class}
\alias{ProfileVar}
\title{NLME Profile variable}
\arguments{
\item{effectName}{Name of fixed effect}

\item{initialvalue}{Initial value}

\item{pertubateValues}{Values to perturbate by (either delta or percentage)}
}
\description{
Class initializer for an NLME profile perturbation variable
}
\examples{
ProfileVar("tvV", 9.95, "-2,-1,0,1,2")
}
\keyword{internal}
