% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomEffectsMethods.R
\name{randomBlockStatement}
\alias{randomBlockStatement}
\title{Returns random block statement}
\usage{
randomBlockStatement(.Object)
}
\arguments{
\item{.Object}{PK/PD model}
}
\description{
Returns random block statement
}
\keyword{internal}
