% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomEffectsMethods.R
\name{randomOccasionalBlockStatement}
\alias{randomOccasionalBlockStatement}
\title{Returns occasional random block statement}
\usage{
randomOccasionalBlockStatement(.Object)
}
\arguments{
\item{.Object}{PK/PD model}
}
\description{
Returns occasional random block statement
}
\keyword{internal}
