% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{residualEffect<-}
\alias{residualEffect<-}
\title{Sets residual effect attributes}
\usage{
residualEffect(.Object, effectName) <- value
}
\arguments{
\item{.Object}{PK/PD model object}

\item{effectName}{Effect to lookup/set attributes for}

\item{value}{A value to be set}
}
\description{
Sets residual effect attributes
}
\keyword{internal}
