% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterNames.R
\name{residualEffectNames}
\alias{residualEffectNames}
\title{Return residual effect terms available in model}
\usage{
residualEffectNames(model)
}
\arguments{
\item{model}{Object of class \code{NlmePmlModel}}
}
\value{
Character vector of residual effect names
}
\description{
Use to return character vector of residual effect names in model object
}
\examples{

model <- pkemaxmodel(columnMap = FALSE)
residualEffectNames(model)

}
