% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveModel.R
\name{saveModel}
\alias{saveModel}
\title{Save model object to .rda file}
\usage{
saveModel(model, engine = NULL, host = NULL)
}
\arguments{
\item{model}{NlmePmlModel object}

\item{engine}{Optional engine parameters to save with model}

\item{host}{Optional host to save with model}
}
\value{
\code{NULL}
}
\description{
Saves the model, engine, and host objects to a single \code{model.rda} file
in model's working directory. If no working directory exists it will be
created by default.
}
\examples{
\donttest{
model <- emaxmodel(
  checkBaseline = TRUE,
  checkFractional = TRUE,
  checkInhibitory = TRUE,
  data = pkpdData,
  ID = "ID",
  C = "CObs",
  EObs = "EObs",
  workingDir = tempdir()
)

saveModel(model)
}

}
\keyword{internal}
