% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveUpdatedMetamodel.R
\name{saveUpdatedMetamodel}
\alias{saveUpdatedMetamodel}
\title{workhorse for change_ThetasMmdlin RsNLME.ModelBuilder}
\usage{
saveUpdatedMetamodel(
  mmdl_withComments,
  mmdl_model,
  resmodel,
  metamodelFile,
  updateModel = TRUE,
  updatedBasedOn = "",
  updatedEstArgsBlock = NULL,
  updatedTablesBlock = NULL
)
}
\arguments{
\item{mmdl_withComments}{the metamodel text to be substituted}

\item{mmdl_model}{the old model from mmdl_withComments}

\item{resmodel}{the model to be pasted}

\item{metamodelFile}{the name of initial file with metamodel
to be overwritten}
}
\value{
text of new metamodel
}
\description{
workhorse for change_ThetasMmdlin RsNLME.ModelBuilder
}
\keyword{internal}
