% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterNames.R
\name{secondaryParameterNames}
\alias{secondaryParameterNames}
\title{Get secondary parameter names}
\usage{
secondaryParameterNames(model)
}
\arguments{
\item{model}{Object of class \code{NlmePmlModel}}
}
\value{
Character vector of secondary parameter names defined in model
}
\description{
Returns character vector of secondary parameter names for model object.
}
\examples{
model <- pkemaxmodel(columnMap = FALSE)
secondaryparms <- secondaryParameterNames(model)

}
