% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SortColumns.R
\docType{class}
\name{SortColumns-class}
\alias{SortColumns-class}
\alias{SortColumns}
\title{Class initializer for SortColumns}
\arguments{
\item{...}{Names of input data columns
(up to 5 for individual models and no limit for population models)
used to sort the input data and model outputs. Can be supplied
as either a single string or  a vector of strings.}
}
\description{
Class represents an NLME sort columns object
}
\examples{
# The following two setups are equivalent
sortColumnSetUp <- SortColumns("Country,City")
sortColumnSetUp <- SortColumns(c("Country","City"))

# The following two setups are equivalent
sortColumnSetUp <- SortColumns("Sort1 Sort2", "Sort3")
sortColumnSetUp <- SortColumns(c("Sort1, Sort2 "), "Sort3 ")
}
\keyword{internal}
