% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StepwiseParams.R
\docType{class}
\name{StepwiseParams-class}
\alias{StepwiseParams-class}
\alias{StepwiseParams}
\title{Class initializer for NLME StepwiseParams}
\arguments{
\item{addPValue}{Numeric. Threshold for adding a covariate effect}

\item{removePValue}{Numeric. Threshold for removing a covariate effect}

\item{method}{-2LL|AIC|BIC; could be abbreviated.}
}
\description{
Class represents an NLME Stepwise search parameters
}
\examples{
StepwiseParams(0.001, 0.001, "BIC")
}
\keyword{internal}
