% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{structuralParam<-}
\alias{structuralParam<-}
\title{Method to set structural parameter attributes}
\usage{
structuralParam(.Object, parameterNames) <- value
}
\arguments{
\item{.Object}{Model with the parameter}

\item{parameterNames}{The names of structural parameters}

\item{value}{A value of the fixed effect to be set}
}
\description{
Method to set structural parameter attributes
}
\examples{
\donttest{
structuralParam(model, "Cl") <- c(style = LogNormal, initialValue = "0.75")

structuralParam(model, "Cl2") <- c(style = Custom, code = "stparm(V=10^(tvlog10V + nlog10V))")
}
}
\keyword{internal}
