% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterNames.R
\name{structuralParameterNames}
\alias{structuralParameterNames}
\title{Get structural parameter names}
\usage{
structuralParameterNames(model, omitEmpties = TRUE)
}
\arguments{
\item{model}{Object of class \code{NlmePmlModel}}

\item{omitEmpties}{Set to \code{TRUE} to omit empty names}
}
\value{
Character vector of structural parameter names defined in model
}
\description{
Returns character vector of structural parameter names for model object.
}
\examples{
model <- pkemaxmodel(columnMap = FALSE)
stparms <- structuralParameterNames(model)

}
