% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/built_in_models.r
\name{textualmodel}
\alias{textualmodel}
\title{Create a textual model object}
\usage{
textualmodel(modelName = "", workingDir = "", data, mdl = NULL)
}
\arguments{
\item{modelName}{Model name to create subdirectory for model output in current working directory.}

\item{workingDir}{Working directory to run the model. Current working directory will be used if \code{workingDir} not specified.}

\item{data}{Input dataset}

\item{mdl}{File path specifying location of test.mdl file}
}
\value{
\code{NlmePmlModel} object
}
\description{
Use to create an empty model object and optionally supply location of .mdl file to initialize model with PML statements.
}
\examples{
\donttest{
model <- textualmodel(data = pkData)
}

}
