% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{updateErrorModel}
\alias{updateErrorModel}
\title{Updates error model}
\usage{
updateErrorModel(.Object)
}
\description{
Updates error model, recreating some structural parameters
that depend on how residual effects are used
}
\keyword{internal}
