% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeDefaultFiles.R
\name{writeDefaultFiles}
\alias{writeDefaultFiles}
\title{Writes out data/column mapping and engine parameter files for the engine}
\usage{
writeDefaultFiles(model, dataset, simParams, Tables, sortColumns = NULL)
}
\arguments{
\item{model}{Model object}

\item{dataset}{Dataset to run simulation/fitting on}

\item{simParams}{Simulation Parameters (simulation tables are written to coldef)}

\item{Tables}{NlmeTableDef class object(s)}

\item{sortColumns}{SortColumns class object used for proper id mapping during
individual modeling}
}
\value{
\code{NULL}
}
\description{
Writes out data/column mapping and engine parameter files for the engine
}
\keyword{internal}
