% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NlmeParallelHost.R
\name{print.NlmeParallelHost}
\alias{print.NlmeParallelHost}
\title{Print method for NlmeParallelHost class}
\usage{
\method{print}{NlmeParallelHost}(x, ...)
}
\arguments{
\item{x}{An NlmeParallelHost object.}

\item{...}{Additional arguments passed to the print function.}
}
\value{
\code{NULL}
}
\description{
This method prints the information of an NlmeParallelHost object.
}
\examples{
host <- NlmeParallelHost(
  sharedDirectory = "~/shared/",
  installationDirectory = "~/nlme/",
  hostName = "my_host",
  machineName = "192.168.1.100",
  hostType = "RHEL8",
  numCores = 8,
  isLocal = FALSE,
  rLocation = "/usr/bin/R",
  scriptPath = "/path/to/script.R",
  userAuthentication = NlmeUserAuthentication(userName = "myuser", userPassword = "mypassword"),
  parallelMethod = NlmeParallelMethod("SGE_MPI")
)

print(host)

}
\keyword{internal}
